Refactor Bliss::SpeechSegment and CorpusDescriptionParser to use Bliss::Orthography instead of std::string for orthographies
