+++ CMakeLists.txt
@@ -25,7 +25,7 @@ set(ARCH_DESCRIPTION
     "${SYSTEM_NAME}-${CMAKE_HOST_SYSTEM_PROCESSOR}-${BUILD_TYPE}"
 )
 if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
-    set(CMAKE_INSTALL_PREFIX
+    set_property(CACHE CMAKE_INSTALL_PREFIX PROPERTY VALUE
         "${CMAKE_CURRENT_SOURCE_DIR}/arch/${ARCH_DESCRIPTION}/"
     )
 endif()
