+++ src/Speech/AccuracyFsaBuilder.cc
@@ -16,7 +16,7 @@
 #include <Am/Utilities.hh>
 #include <Bliss/Evaluation.hh>
 #include <Bliss/Lexicon.hh>
-#include <Bliss/Orthography.hh>
+#include <Bliss/OrthographicParser.hh>
 #include <Core/StringUtilities.hh>
 #include <Fsa/Arithmetic.hh>
 #include <Fsa/Basic.hh>
