+++ src/Speech/LatticeExtractor.cc
@@ -15,7 +15,7 @@
 #include "LatticeExtractor.hh"
 #include <Am/Module.hh>
 #include <Am/Utilities.hh>
-#include <Bliss/Orthography.hh>
+#include <Bliss/OrthographicParser.hh>
 #include <Core/Hash.hh>
 #include <Core/Vector.hh>
 #include <Fsa/Arithmetic.hh>
